requirements.txt
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
click==8.1.8
colorama==0.4.6
Flask==2.2.5
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.6
joblib==1.3.2
llvmlite==0.39.1
MarkupSafe==2.1.5
numba==0.56.4
numpy==1.21.6
pandas==1.3.5
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.17.0
threadpoolctl==3.1.0
tqdm==4.67.1
typing_extensions==4.7.1
umap-learn==0.5.7
Werkzeug==2.2.3
wincertstore==0.2
zipp==3.15.0
